move assert class
